SemiparametricAUC.jl Commands

SemiparametricAUC.calculate_auc — Method.

calculate_auc(ya, yb)

This function takes two DataArrays ya and yb and calculates variance of predicted AUC, logit of predicted AUC, and variance of logit of predicted AUC responses passed.

source

SemiparametricAUC.semiparametricAUC — Method.

SemiparametricAUC.semiparametricAUC(model_formula = y ~ x1, treatment_group = :group, data = fasd)

This function is used to fit semiparametric AUC regression model specified by giving a formula object of response and covariates and a separate argument of treatment group. It will convert variables other than response into factors, estimate model parameters, and display results.

source

SemiparametricAUC.calculate_auc_simulation — Method.

calculate_auc_simulation(ya, yb)

This function takes two DataArrays ya and yb and calculates variance of predicted AUC, logit of predicted AUC, and variance of logit of predicted AUC responses passed.

source

SemiparametricAUC.simulate_one_predictor — Method.

simulate_one_predictor(iter, m, p)

It asks for number of iterations to be run, number of observations in treatment and control groups for the simulation of Semiparametric AUC regression adjusting for one discrete covariate. In this simulation, true model parameters are as follows: β0 = 0.15, β1 = 0.50, β2 = 1.

source