SemiparametricAUC.jl Commands

SemiparametricAUC.calculate_auc — Method.

calculate_auc(ya, yb)

"Calcualtes AUC related estimates"

This function takes two DataArray arguments ya and yb and calculates variance of predicted AUC, logit of predicted AUC, and variance of logit of predicted AUC responses passed.

source

SemiparametricAUC.semiparametricAUC — Method.

semiparametricAUC(model_formula, treatment_group, data)

This function is used to fit semiparametric AUC regression model specified by giving a formula object of response and covariates and a separate argument of treatment group. It will convert variables other than response into factors, estimate model parameters, and display results.

This function takes a formula object model_formula with response and covariates such as {response ~ x1 + x2}, group argument treatment_group, which is treatment group for which a comparision is to be made, and a data argument data which is a DataFrame object that contains variables needed for the analysis.

source

SemiparametricAUC.calculate_auc_simulation — Method.

calculate_auc_simulation(ya, yb)

This function takes two DataArray arguments ya and yb and calculates variance of predicted AUC, logit of predicted AUC, and variance of logit of predicted AUC responses passed.

source

SemiparametricAUC.simulate_one_predictor — Function.

simulate_one_predictor(iter, m, p)

It asks for number of iterations iter to be run, number of observations m in treatment and control groups p for the simulation of Semiparametric AUC regression adjusting for one discrete covariate. In this simulation, true model parameters are as follows: β0 = 0.15, β1 = 0.50, β2 = 1.

source

SemiparametricAUC.coefs_table — Method.

coefs_table(mf,lo, up, betass, std_errors)

This function takes a ModelFrame object mf, numeric arguments lo, up, betass estimates and std_errors (beta's standard errors, returns a table with model estimates, 95% CI, and p-values.

source